+++ app/src/main/java/id/co/qualitas/qubes/fragment/aspp/RouteCustomerFragment.java
@@ -129,7 +129,7 @@ public class RouteCustomerFragment extends BaseFragment implements LocationListe
             public void onRefresh() {
                 if (SessionManagerQubes.getStartDay() == 0) {
                     requestData();
-                }else{
+                } else {
                     setToast("Sudah start visit");
                 }
                 swipeLayout.setRefreshing(false);
@@ -255,49 +255,60 @@ public class RouteCustomerFragment extends BaseFragment implements LocationListe
                         SessionManagerQubes.setStartDay(0);
                     }
                     Customer[] param1Array = Helper.ObjectToGSON(result.get("customerNonRoute"), Customer[].class);
-                    Collections.addAll(mListNonRoute, param1Array);
-                    database.deleteMasterNonRouteCustomer();
-                    database.deleteMasterNonRouteCustomerPromotion();
-                    database.deleteMasterNonRouteCustomerDct();
+                    if (param1Array != null) {
+                        Collections.addAll(mListNonRoute, param1Array);
+                        database.deleteMasterNonRouteCustomer();
+                        database.deleteMasterNonRouteCustomerPromotion();
+                        database.deleteMasterNonRouteCustomerDct();
+                    }
 
                     for (Customer param : mListNonRoute) {
                         List<Promotion> arrayList = new ArrayList<>();
                         Promotion[] matArray = Helper.ObjectToGSON(param.getPromoList(), Promotion[].class);
-                        Collections.addAll(arrayList, matArray);
-                        param.setPromoList(arrayList);
+                        if (matArray != null) {
+                            Collections.addAll(arrayList, matArray);
+                            param.setPromoList(arrayList);
+                        }
+
+                        List<Material> arrayDctList = new ArrayList<>();
+                        Material[] dctArray = Helper.ObjectToGSON(param.getDctList(), Material[].class);
+                        if (dctArray != null) {
+                            Collections.addAll(arrayDctList, dctArray);
+                        }
+                        param.setDctList(arrayDctList);
 
                         int idHeader = database.addNonRouteCustomer(param, user.getUsername());
                         for (Promotion mat : arrayList) {
                             database.addNonRouteCustomerPromotion(mat, String.valueOf(idHeader), user.getUsername());
                         }
-
-                        List<Material> arrayDctList = new ArrayList<>();
-                        Material[] dctArray = Helper.ObjectToGSON(param.getDctList(), Material[].class);
-                        Collections.addAll(arrayDctList, dctArray);
-                        param.setDctList(arrayDctList);
-
                         for (Material mat : arrayDctList) {
                             database.addNonRouteCustomerDct(mat, String.valueOf(idHeader), user.getUsername(), param.getId());
                         }
                     }
 
                     Customer[] paramArray = Helper.ObjectToGSON(result.get("todayCustomer"), Customer[].class);
-                    Collections.addAll(mList, paramArray);
-                    database.deleteCustomer();
-                    database.deleteCustomerPromotion();
-                    database.deleteCustomerDct();
-                    database.deleteVisitSalesman();
-                    database.deleteNoo();
+                    if (paramArray != null) {
+                        Collections.addAll(mList, paramArray);
+                        database.deleteCustomer();
+                        database.deleteCustomerPromotion();
+                        database.deleteCustomerDct();
+                        database.deleteVisitSalesman();
+                        database.deleteNoo();
+                    }
 
                     for (Customer param : mList) {
                         List<Promotion> arrayList = new ArrayList<>();
                         Promotion[] matArray = Helper.ObjectToGSON(param.getPromoList(), Promotion[].class);
-                        Collections.addAll(arrayList, matArray);
+                        if (matArray != null) {
+                            Collections.addAll(arrayList, matArray);
+                        }
                         param.setPromoList(arrayList);
 
                         List<Material> arrayDctList = new ArrayList<>();
                         Material[] dctArray = Helper.ObjectToGSON(param.getDctList(), Material[].class);
-                        Collections.addAll(arrayDctList, dctArray);
+                        if (dctArray != null) {
+                            Collections.addAll(arrayDctList, dctArray);
+                        }
                         param.setDctList(arrayDctList);
 
                         int idHeader = database.addCustomer(param, user.getUsername());
